refactor: extract the crate-size check into its own module
Move `checkCratesPackageSize` and its cargo-package helpers out of
`check.ts` into a dedicated `check-crate-size.ts`, so the test file
mirrors the file under test and the module holds one check.

The unit test now calls `checkCratesPackageSize` directly with
hand-built packages, mocking only the third-party `cargo` subprocess
at the `spawnSync` boundary — no git repo or config loader. The
integration test continues to drive the check through `runChecks`,
mocking only the same third-party subprocess.

No behaviour change.
