+++ src/check-crate-size.test.ts
@@ -1,36 +1,42 @@
 /**
- * Unit tests for `runChecks`' crate-size pre-merge check (#362).
+ * Unit tests for `checkCratesPackageSize` (#362).
  *
- * The behavioural contract is also exercised in
- * `test/integration/check-crate-size.integration.test.ts`. These cases
- * own coverage: the integration config is excluded from
- * `test:unit:coverage` per `vitest.config.ts`, so every branch of the
- * size check needs a unit case here — the cargo subprocess is faked at
- * the `spawnSync` boundary so the assertions are deterministic and need
- * no Rust toolchain.
+ * The file under test is `check-crate-size.ts`. Its only dependency
+ * outside itself is the `cargo package` subprocess, mocked here at the
+ * `spawnSync` boundary — so these cases call the check function
+ * directly with hand-built packages, with no git repo, config loader,
+ * or Rust toolchain involved. They own branch coverage of the module.
  *
- * Each case stands up a throwaway git repo (cheap — `runChecks` shells
- * out to `git ls-files`) and drives `runChecks` against a hand-built
- * `putitoutthere.toml` plus the manifests the check reads.
+ * The end-to-end path through `runChecks` (real config loader, real
+ * check dispatch) is covered by
+ * `test/integration/check-crate-size.integration.test.ts`.
  */
 
 import type * as ChildProcess from 'node:child_process';
-import { execFileSync, spawnSync } from 'node:child_process';
-import { mkdirSync, mkdtempSync, rmSync, writeFileSync } from 'node:fs';
-import { tmpdir } from 'node:os';
-import { join } from 'node:path';
-import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
+import { spawnSync } from 'node:child_process';
+import { afterEach, describe, expect, it, vi } from 'vitest';
 
 vi.mock('node:child_process', async (orig) => {
   const actual = await orig<typeof ChildProcess>();
-  // git stays real (`runChecks` walks `git ls-files`); only the cargo
-  // subprocess the size check shells out to is faked.
-  return { ...actual, spawnSync: vi.fn(actual.spawnSync) };
+  return { ...actual, spawnSync: vi.fn() };
 });
 
-import { runChecks } from './check.js';
+import { checkCratesPackageSize } from './check-crate-size.js';
+import type { Package } from './config.js';
 
 const spawnMock = vi.mocked(spawnSync);
+
+function makePkg(kind: Package['kind'], name = `${kind}-pkg`): Package {
+  return {
+    name,
+    kind,
+    path: 'packages/pkg',
+    globs: ['packages/pkg/**'],
+    depends_on: [],
+    first_version: '0.1.0',
+    tag_format: '{name}-v{version}',
+  } as unknown as Package;
+}
 
 /** A `cargo package` run that exited 0 and reported `compressed` size. */
 function cargoPackaged(compressed: string): SpawnSyncReturnsString {
@@ -45,91 +51,28 @@ function cargoExit0(stderr: string): SpawnSyncReturnsString {
 
 type SpawnSyncReturnsString = ReturnType<typeof spawnSync> & { stderr: string };
 
-let cwd: string;
-
-function git(args: string[]): void {
-  execFileSync('git', args, { cwd, stdio: ['ignore', 'pipe', 'pipe'] });
-}
-
-function write(rel: string, body: string): void {
-  const full = join(cwd, rel);
-  mkdirSync(join(full, '..'), { recursive: true });
-  writeFileSync(full, body, 'utf8');
-}
-
-function commit(): void {
-  git(['add', '-A']);
-  git(['commit', '-q', '-m', 'snapshot']);
-}
-
-/** Seed a committed repo with one well-formed `kind = "crates"` package. */
-function seedCratesRepo(): void {
-  write(
-    'putitoutthere.toml',
-    `
-[putitoutthere]
-version = 1
-
-[[package]]
-name  = "rust-lib"
-kind  = "crates"
-path  = "packages/rs"
-globs = ["packages/rs/**"]
-`,
-  );
-  write(
-    'packages/rs/Cargo.toml',
-    `
-[package]
-name = "rust-lib"
-version = "0.1.0"
-description = "a crate"
-license = "MIT"
-`,
-  );
-  write('packages/rs/src/lib.rs', '');
-  commit();
-}
-
-function hasSizeFinding(messages: { message: string }[]): boolean {
-  return messages.some((f) => /PIOT_CRATES_PACKAGE_TOO_LARGE/.test(f.message));
-}
-
-beforeEach(() => {
-  cwd = mkdtempSync(join(tmpdir(), 'piot-crate-size-unit-'));
-  git(['init', '-q', '-b', 'main']);
-  git(['config', 'user.email', 't@example.com']);
-  git(['config', 'user.name', 't']);
-  git(['config', 'commit.gpgsign', 'false']);
-  spawnMock.mockReturnValue(cargoPackaged('8.9KiB'));
-});
-
 afterEach(() => {
   spawnMock.mockReset();
-  rmSync(cwd, { recursive: true, force: true });
 });
 
-describe('runChecks: crate-size pre-merge check (#362)', () => {
+describe('checkCratesPackageSize (#362)', () => {
   it("flags a crates package whose .crate exceeds crates.io's 10 MiB limit", () => {
-    seedCratesRepo();
     spawnMock.mockReturnValue(cargoPackaged('133.6MiB'));
-    const findings = runChecks({ cwd });
-    const hit = findings.find((f) => /PIOT_CRATES_PACKAGE_TOO_LARGE/.test(f.message));
-    expect(hit).toBeDefined();
-    expect(hit!.package).toBe('rust-lib');
-    expect(hit!.message).toMatch(/133\.6 MiB/);
-    expect(hit!.message).toMatch(/10\.0 MiB|10485760/);
-    expect(hit!.message).toMatch(/413 Payload Too Large/);
+    const findings = checkCratesPackageSize([makePkg('crates', 'rust-lib')]);
+    expect(findings).toHaveLength(1);
+    expect(findings[0]!.package).toBe('rust-lib');
+    expect(findings[0]!.message).toMatch(/PIOT_CRATES_PACKAGE_TOO_LARGE/);
+    expect(findings[0]!.message).toMatch(/133\.6 MiB/);
+    expect(findings[0]!.message).toMatch(/10\.0 MiB|10485760/);
+    expect(findings[0]!.message).toMatch(/413 Payload Too Large/);
   });
 
   it('does not flag a crates package whose .crate is within the limit', () => {
-    seedCratesRepo();
     spawnMock.mockReturnValue(cargoPackaged('2.0MiB'));
-    expect(hasSizeFinding(runChecks({ cwd }))).toBe(false);
+    expect(checkCratesPackageSize([makePkg('crates')])).toEqual([]);
   });
 
-  it('does not flag when cargo is absent (spawnSync reports an error)', () => {
-    seedCratesRepo();
+  it('returns nothing when cargo is absent (spawnSync reports an error)', () => {
     spawnMock.mockReturnValue({
       pid: 0,
       output: [],
@@ -139,11 +82,10 @@ describe('runChecks: crate-size pre-merge check (#362)', () => {
       signal: null,
       error: Object.assign(new Error('spawnSync cargo ENOENT'), { code: 'ENOENT' }),
     });
-    expect(hasSizeFinding(runChecks({ cwd }))).toBe(false);
+    expect(checkCratesPackageSize([makePkg('crates')])).toEqual([]);
   });
 
-  it('does not flag when cargo package exits non-zero', () => {
-    seedCratesRepo();
+  it('returns nothing when cargo package exits non-zero', () => {
     spawnMock.mockReturnValue({
       pid: 1,
       output: ['', 'error: failed to parse manifest\n'],
@@ -152,56 +94,30 @@ describe('runChecks: crate-size pre-merge check (#362)', () => {
       status: 101,
       signal: null,
     });
-    expect(hasSizeFinding(runChecks({ cwd }))).toBe(false);
+    expect(checkCratesPackageSize([makePkg('crates')])).toEqual([]);
   });
 
-  it('does not flag when spawnSync itself throws', () => {
-    seedCratesRepo();
+  it('returns nothing when spawnSync itself throws', () => {
     spawnMock.mockImplementation(() => {
       throw new Error('spawn failed');
     });
-    expect(hasSizeFinding(runChecks({ cwd }))).toBe(false);
+    expect(checkCratesPackageSize([makePkg('crates')])).toEqual([]);
   });
 
-  it('does not flag when cargo output carries no Packaged size line', () => {
-    seedCratesRepo();
+  it('returns nothing when cargo output carries no Packaged size line', () => {
     spawnMock.mockReturnValue(cargoExit0('   Compiling rust-lib v0.1.0\n'));
-    expect(hasSizeFinding(runChecks({ cwd }))).toBe(false);
+    expect(checkCratesPackageSize([makePkg('crates')])).toEqual([]);
   });
 
-  it('does not flag when cargo reports an unrecognised size unit', () => {
-    seedCratesRepo();
+  it('returns nothing when cargo reports an unrecognised size unit', () => {
     spawnMock.mockReturnValue(
       cargoExit0('    Packaged 7 files, 24.0KiB (5.0ZB compressed)\n'),
     );
-    expect(hasSizeFinding(runChecks({ cwd }))).toBe(false);
+    expect(checkCratesPackageSize([makePkg('crates')])).toEqual([]);
   });
 
-  it('skips non-crates packages — cargo is never invoked for them', () => {
-    write(
-      'putitoutthere.toml',
-      `
-[putitoutthere]
-version = 1
-
-[[package]]
-name  = "js-lib"
-kind  = "npm"
-path  = "packages/js"
-globs = ["packages/js/**"]
-`,
-    );
-    write(
-      'packages/js/package.json',
-      JSON.stringify({
-        name: 'js-lib',
-        version: '0.0.0',
-        repository: { type: 'git', url: 'git+https://github.com/x/y.git' },
-      }),
-    );
-    write('packages/js/index.ts', 'x');
-    commit();
-    expect(hasSizeFinding(runChecks({ cwd }))).toBe(false);
+  it('skips non-crates packages without invoking cargo', () => {
+    expect(checkCratesPackageSize([makePkg('npm'), makePkg('pypi')])).toEqual([]);
     expect(spawnMock).not.toHaveBeenCalled();
   });
 });
