+++ src/check.ts
@@ -23,13 +23,14 @@
  * Issue #319.
  */
 
-import { execFileSync, spawnSync, type SpawnSyncReturns } from 'node:child_process';
+import { execFileSync } from 'node:child_process';
 import { existsSync, readFileSync, statSync } from 'node:fs';
 import { dirname, isAbsolute, join, resolve } from 'node:path';
 
 import { parse as parseToml } from 'smol-toml';
 
 import { assertNoCycles } from './cascade.js';
+import { checkCratesPackageSize } from './check-crate-size.js';
 import { loadConfig, type Package } from './config.js';
 import { ErrorCodes } from './error-codes.js';
 import { matchesAny } from './glob.js';
@@ -103,7 +104,7 @@ export function runChecks(opts: CheckOptions): CheckFinding[] {
   checkTagTemplateCollisions(packages, findings);
   checkNpmRepository(packages, findings);
   checkCratesPackageMetadata(packages, findings);
-  checkCratesPackageSize(packages, findings);
+  findings.push(...checkCratesPackageSize(packages));
   checkPyprojectAndBundleCli(packages, cwd, findings);
   checkPypiVersion(packages, findings);
   checkPyprojectShapeFindings(packages, findings);
@@ -244,83 +245,6 @@ function checkCratesPackageMetadata(
       message: `[${ErrorCodes.CRATES_MISSING_METADATA}] ${f.cargoTomlPath} missing required Cargo.toml [package] field(s): ${f.missing.join(', ')}. crates.io rejects the publish after cargo's verification build.`,
     });
   }
-}
-
-// crates.io rejects any `.crate` upload over this size with
-// `413 Payload Too Large`. cargo surfaces the rejection only after
-// `cargo publish` has run the verification build and compiled every
-// transitive dep; `cargo package` reproduces the identical tarball at
-// PR time in seconds.
-const CRATES_MAX_UPLOAD_BYTES = 10 * 1024 * 1024;
-
-const CARGO_SIZE_UNIT_BYTES: Record<string, number> = {
-  B: 1,
-  KiB: 1024,
-  MiB: 1024 ** 2,
-  GiB: 1024 ** 3,
-  TiB: 1024 ** 4,
-};
-
-function checkCratesPackageSize(
-  packages: readonly Package[],
-  findings: CheckFinding[],
-): void {
-  for (const p of packages) {
-    if (p.kind !== 'crates') continue;
-    const cargoTomlPath = join(p.path, 'Cargo.toml');
-    const compressedBytes = packagedCrateBytes(cargoTomlPath);
-    if (compressedBytes === null) continue;
-    if (compressedBytes > CRATES_MAX_UPLOAD_BYTES) {
-      findings.push({
-        package: p.name,
-        message: `[${ErrorCodes.CRATES_PACKAGE_TOO_LARGE}] ${cargoTomlPath}: cargo package produces a ${formatMiB(compressedBytes)} .crate, over crates.io's ${formatMiB(CRATES_MAX_UPLOAD_BYTES)} (${CRATES_MAX_UPLOAD_BYTES}-byte) upload limit. cargo publish would fail with 413 Payload Too Large mid-release; a tracked symlink into a build directory or a missing [package].exclude is the usual cause.`,
-      });
-    }
-  }
-}
-
-/**
- * Compressed size in bytes of the `.crate` `cargo package` would
- * upload, or null when it can't be determined — cargo is absent, cargo
- * rejected the manifest, or its output carried no size line. Null means
- * "can't verify": the check skips rather than inventing a size finding,
- * mirroring the null-means-skip shape `listTrackedFiles` uses outside a
- * git repo.
- */
-function packagedCrateBytes(cargoTomlPath: string): number | null {
-  let result: SpawnSyncReturns<string>;
-  try {
-    result = spawnSync(
-      'cargo',
-      ['package', '--no-verify', '--allow-dirty', '--manifest-path', cargoTomlPath],
-      { encoding: 'utf8', stdio: ['ignore', 'pipe', 'pipe'] },
-    );
-  } catch {
-    return null;
-  }
-  if (result.error !== undefined) return null;
-  if (result.status !== 0) return null;
-  return parseCargoCompressedBytes(result.stderr);
-}
-
-/**
- * Pull the compressed `.crate` size out of `cargo package`'s summary
- * line, which cargo writes to stderr:
- *   `    Packaged 7 files, 24.0KiB (8.9KiB compressed)`
- */
-function parseCargoCompressedBytes(cargoStderr: string): number | null {
-  const m = /Packaged\b[^\n]*?\(\s*(\d+(?:\.\d+)?)\s*([A-Za-z]+)\s+compressed\s*\)/.exec(
-    cargoStderr,
-  );
-  if (m === null) return null;
-  const [, rawValue, rawUnit] = m as unknown as [string, string, string];
-  const unit = CARGO_SIZE_UNIT_BYTES[rawUnit];
-  if (unit === undefined) return null;
-  return Math.round(Number.parseFloat(rawValue) * unit);
-}
-
-function formatMiB(bytes: number): string {
-  return `${(bytes / (1024 * 1024)).toFixed(1)} MiB`;
 }
 
 function checkPyprojectAndBundleCli(
